tests/wpt/tests/signed-exchange/fallback-to-another-sxg.tentative.html

<!DOCTYPE html>
<title>Fallback redirect to another SignedHTTPExchange</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="./resources/sxg-util.js"></script>
<body>
<script>
promise_test(async (t) => {
  const sxgUrl = get_host_info().HTTPS_ORIGIN + '/signed-exchange/resources/sxg/fallback-to-another-sxg.sxg';
  const message = await openSXGInIframeAndWaitForMessage(t, sxgUrl);
  assert_equals(message.location, innerURLOrigin() + '/signed-exchange/resources/inner-url.html');
  assert_false(message.is_fallback);
}, 'Fallback redirect to another SignedHTTPExchange');

</script>
</body>
